requirements.txt
boto3
botocore
certifi
chardet
dj-database-url
Django==3.1
django-crispy-forms
django-heroku
django-storages
docutils
gunicorn
idna
jmespath
Pillow
psycopg2
python-dateutil
pytz
requests
s3transfer
six
urllib3
whitenoise
django-cleanup
django-address
django-phone-field
django-countries
django-localflavor
